// lib/faust-types.ts
export interface ImpactHeroFields {
  heroSubtitle: string;
  heroTitle1: string;
  heroTitleItalic: string;
  heroTitle2: string;
  heroTitleEnd: string;
  heroVision: string;
  heroMission: string;
  heroBadgeNum: string;
  heroQuote: string;
  heroImage: {
    node: {
      sourceUrl: string;
      altText: string;
    };
  };
}

export interface ImpactHero {
  title: string;
  impactHeroFields: ImpactHeroFields;
}

export interface ImpactHeroData {
  impactHeroes: {
    nodes: ImpactHero[];
  };
}